moved various dev pages
