+++ pages/dev-pages/components.tsx
@@ -6,14 +6,16 @@ import { NotificationsWidget } from "@/components/NotificationsWidget";
 import { Notification } from "@/src/types/Notification";
 import { GameWidget } from "@/components/GameWidget";
 import Head from "next/head";
-import NotificationBubble from "../components/NotificationBubble";
-import SearchBar from "../components/SearchBar";
-import SideNav from "../components/SideNav";
+import NotificationBubble from "../../components/NotificationBubble";
+import SearchBar from "../../components/SearchBar";
+import SideNav from "../../components/SideNav";
 import { GamesInterests } from "@/components/GamesInterests";
 import { HonorOfConduct } from "@/components/HonorOfConduct";
 import PartyOption from "@/components/PartyOptions";
 import { ProfilePreview } from "@/components/ProfilePreview";
 import ProfileCard from "@/components/ProfileCard";
+import Link from "next/link";
+import React from "react";
 
 export default function Components() {
   const notifications: Notification[] = [
@@ -43,31 +45,30 @@ export default function Components() {
     }
   ];
   return (
-    <>
-      <Head>
-        <title>Project Design Thriving</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main>
-        <LoginStatus />
-        <SideNav />
-        <SearchBar />
-        <AchievementsWidget />
-        <InterestsWidget />
-        <ProfilePreview />
-        <CustomizeDashboard />
-        <NotificationsWidget notifications={notifications} />
-        {notifications.map((notification, index) => (
-          <NotificationBubble key={index} notification={notification} />
-        ))}
-        <GamesInterests />
-        <GameWidget />
-        <HonorOfConduct />
-        <PartyOption />
-        <ProfileCard />
-      </main>
-    </>
+    <main className={"p-8"}>
+      <Link href={"/dev-pages"}>
+        <button>
+          Back
+        </button>
+      </Link>
+
+      <h1 className={"mb-8"}>Various Components</h1>
+
+      <SideNav />
+      <SearchBar />
+      <AchievementsWidget />
+      <InterestsWidget />
+      <ProfilePreview />
+      <CustomizeDashboard />
+      <NotificationsWidget notifications={notifications} />
+      {notifications.map((notification, index) => (
+        <NotificationBubble key={index} notification={notification} />
+      ))}
+      <GamesInterests />
+      <GameWidget />
+      <HonorOfConduct />
+      <PartyOption />
+      <ProfileCard />
+    </main>
   );
 }
