+++ pages/dev-pages/index.tsx
@@ -24,6 +24,26 @@ const DevPages = () => {
         Atoms
       </Button>
     </Link>
+
+    <h3 className={"mt-8"}>Miscellaneous</h3>
+
+    <Link href={"/dev-pages/components"}>
+      <Button type={"secondary"}>See components page</Button>
+    </Link>
+
+    <br/>
+    <br/>
+
+    <Link href={"/dev-pages/chat-demo"}>
+      <Button type={"secondary"}>See chat demo</Button>
+    </Link>
+
+    <br/>
+    <br/>
+
+    <Link href={"/dev-pages/party-finder/test"}>
+      <Button type={"secondary"}>See Party Finder Backend Tests</Button>
+    </Link>
   </main>;
 };
